Crate parachain_runtime

The Robonomics runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants
A set of constant values used in substrate runtime.

Structs

AdjustmentVariable

Ancestry

BasicDeposit

BlockHashCount

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

Burn

CandidacyBond

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

DataDepositPerByte

DesiredMembers

DesiredRunnersUp

ElectionsPhragmenModuleId

ExistentialDeposit

FieldDeposit

GenesisConfig

MaxAdditionalFields

MaxLocks

MaxRegistrars

MaxScheduledPerBlock

MaxSubAccounts

MaximumReasonLength

MaximumSchedulerWeight

MinimumMultiplier

MinimumPeriod

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PointsLimit

ProposalBond

ProposalBondMinimum

RelayChainOrigin

RococoLocation

RococoNetwork

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

SpendPeriod

SubAccountDeposit

TargetBlockFullness

TermDuration

TipCountdown

TipFindersFee

TipReportDepositBase

TotalBandwidth

TransactionByteFee

TreasuryModuleId

Version

VotingBondBase

VotingBondFactor

WeightLimit

XcmConfig

Enums

Call

Event

OriginCaller

Constants

VERSION
Standalone runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version
The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap
Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Definitions

Address
The address format for describing accounts.

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

Bounties

CheckedExtrinsic
Extrinsic type that has already been checked.

Council

CouncilConfig

Datalog

DatalogXcm

DigitalTwin

Elections

ElectionsConfig

Executive
Executive: handles dispatch to the various modules.

Header
Block header type as expected by this runtime.

Identity

Launch

LaunchXcm

ParachainInfo

ParachainInfoConfig

ParachainSystem

RWS

RandomnessCollectiveFlip

Scheduler

SignedExtra
The SignedExtension to the basic transaction logic.

Sudo

SudoConfig

System

SystemConfig

Timestamp

Tips

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility

XcmHandler